Share telemetry query adapter
