+++ apps/api/package.json
@@ -40,6 +40,7 @@
     "@superlog/net-guard": "workspace:*",
     "@superlog/railway": "workspace:*",
     "@superlog/render": "workspace:*",
+    "@superlog/telemetry-query": "workspace:*",
     "autumn-js": "^1.2.27",
     "better-auth": "^1.6.13",
     "dotenv": "^17.4.2",
